[payload-generator] Promote new impl to index
