+++ packages/graphql-js-operation-payload-generator/src/index.ts
@@ -1,35 +1,26 @@
-import type {
-  ASTKindToNode,
-  DefinitionNode,
-  DocumentNode,
-  FieldNode,
-  FragmentDefinitionNode,
-  OperationDefinitionNode,
-  SelectionSetNode,
-  GraphQLCompositeType,
-  GraphQLObjectType,
-  GraphQLOutputType,
-  GraphQLScalarType,
-  GraphQLSchema,
-} from "graphql";
 import {
-  assertCompositeType,
+  executeSync,
   getNamedType,
-  getNullableType,
+  getOperationAST,
   isAbstractType,
+  isCompositeType,
   isEnumType,
   isListType,
-  isObjectType,
   isScalarType,
-  TypeInfo,
   visit,
-  Visitor,
-  visitWithTypeInfo,
+  DocumentNode,
+  GraphQLResolveInfo,
+  GraphQLSchema,
+  GraphQLAbstractType,
+  GraphQLCompositeType,
+  GraphQLNamedType,
+  GraphQLScalarType,
+  FieldNode,
+  FragmentDefinitionNode,
+  SelectionNode,
+  SelectionSetNode,
 } from "graphql";
-import { Maybe } from "graphql/jsutils/Maybe";
-import invariant from "invariant";
-import deepmerge from "deepmerge";
-
+import { Path, pathToArray } from "graphql/jsutils/Path";
 import {
   createValueResolver,
   DEFAULT_MOCK_RESOLVERS,
@@ -39,6 +30,8 @@ import {
   MockResolvers,
   ValueResolver,
 } from "./vendor/RelayMockPayloadGenerator";
+import invariant from "invariant";
+
 export { MockResolvers } from "./vendor/RelayMockPayloadGenerator";
 
 export interface RequestDescriptor<Node = DocumentNode> {
@@ -54,368 +47,362 @@ export interface OperationDescriptor<
   readonly request: RequestDescriptor<Node>;
 }
 
-type NodeWithMockData = { userMockData: unknown };
-
-type VisitorWithMockData = Visitor<
-  ASTKindToNode & {
-    Field: Partial<NodeWithMockData>;
-  }
->;
+interface InternalMockData extends MockData {
+  __abstractType?: GraphQLAbstractType;
+}
 
 const TYPENAME_KEY = "__typename";
 
 export function generate(
   operation: OperationDescriptor,
   mockResolvers: MockResolvers | null = DEFAULT_MOCK_RESOLVERS,
-) {
+): { data: MockData } {
   mockResolvers = { ...DEFAULT_MOCK_RESOLVERS, ...mockResolvers };
   const resolveValue = createValueResolver(mockResolvers);
-  const definitions = operation.request.node.definitions;
 
-  const operationDefinitionNode = definitions.find(
-    (def) => def.kind === "OperationDefinition",
-  ) as OperationDefinitionNode | undefined;
-  invariant(operationDefinitionNode, "Expected an operation definition node");
-
-  const result = visitDocumentDefinitionNode(
-    operationDefinitionNode,
-    operation.schema,
-    definitions,
-    mockResolvers,
-    resolveValue,
-  );
-
-  return { data: result };
-}
-
-function isNodeWithMockData(node: any): node is NodeWithMockData {
-  return (
-    node !== undefined &&
-    (node as Partial<NodeWithMockData>).userMockData !== undefined
-  );
-}
-
-/**
- * Transforms a graphql-js AST to mock data.
- *
- * We generate the data upon entering a node and return the data when leaving
- * the node.
- */
-function visitDocumentDefinitionNode(
-  documentDefinitionNode:
-    | OperationDefinitionNode
-    | (FragmentDefinitionNode & Partial<NodeWithMockData>),
-  schema: GraphQLSchema,
-  allDocumentDefinitionNodes: ReadonlyArray<DefinitionNode>,
-  mockResolvers: MockResolvers,
-  resolveValue: ValueResolver,
-): MockData {
-  const typeInfo = new TypeInfo(schema);
-  const visitor: VisitorWithMockData = {
-    OperationDefinition: {
-      /**
-       * Return collected mock data
-       */
-      leave(operationDefinitionNode) {
-        return operationDefinitionNode.selectionSet;
-      },
-    },
-    FragmentDefinition: {
-      /**
-       * Return collected mock data
-       */
-      leave(fragmentDefinitionNode) {
-        return fragmentDefinitionNode.selectionSet;
-      },
-    },
-    InlineFragment: {
-      /**
-       * Return collected mock data
-       */
-      leave(inlineFragmentNode) {
-        return inlineFragmentNode.selectionSet;
-      },
-    },
-    FragmentSpread: {
-      /**
-       * Visit referenced fragment and return collected MockData
-       */
-      leave(fragmentSpreadNode, _key, _parent, _path, ancestors) {
-        // From the parent Field/FragmentDefinition node, get existing user
-        // mock data and make it available to the new visitor.
-        const parentNodeWithMockData = ancestors[ancestors.length - 2];
-        const userMockData = isNodeWithMockData(parentNodeWithMockData)
-          ? parentNodeWithMockData.userMockData
-          : undefined;
-
-        const fragmentDefinitionNode = findFragmentDefinitionNode(
-          allDocumentDefinitionNodes,
-          fragmentSpreadNode.name.value,
-        );
-        return visitDocumentDefinitionNode(
-          { ...fragmentDefinitionNode, userMockData },
-          schema,
-          allDocumentDefinitionNodes,
-          mockResolvers,
-          resolveValue,
-        );
-      },
-    },
-    SelectionSet: {
-      /**
-       * Reduce possible concrete types for abstract type to a single one.
-       */
-      enter(selectionSetNode, _key, parent) {
-        if (isAbstractType(typeInfo.getType())) {
-          return reduceToSingleObjectTypeSelection(
-            selectionSetNode,
-            isNodeWithMockData(parent)
-              ? (parent.userMockData as MockData)[TYPENAME_KEY]
-              : undefined,
-            schema,
-            allDocumentDefinitionNodes,
-            typeInfo,
-          );
-        }
-      },
-      /**
-       * Merge mock data of all selections and return it
-       */
-      leave(selectionSetNode) {
-        const type = typeInfo.getType();
-        const mocksData = (selectionSetNode.selections as unknown[]) as MockData[];
-
-        /**
-         * Always add __typename to object types so it's made available in cases
-         * where an abstract parent type had a field selection for it; in which
-         * case it needs an object type name, not the abstract type's name.
-         */
-        // First search for a schema typename...
-        let typename = mocksData.find(
-          (sel) =>
-            typeof sel[TYPENAME_KEY] === "string" &&
-            sel[TYPENAME_KEY] !== DEFAULT_MOCK_TYPENAME,
-        )?.[TYPENAME_KEY];
-        // ...otherwise use the current type, if it's an object type...
-        if (!typename && isObjectType(type)) {
-          typename = type.name;
-        }
-        // ...or fallback to a DEFAULT_MOCK_TYPENAME selection.
-        if (
-          !typename &&
-          mocksData.some((sel) => sel[TYPENAME_KEY] === DEFAULT_MOCK_TYPENAME)
-        ) {
-          typename = DEFAULT_MOCK_TYPENAME;
-        }
-
-        const mockData: MockData = deepmerge.all(mocksData) as MockData;
-        if (typename) {
-          delete mockData[TYPENAME_KEY];
-          return {
-            [TYPENAME_KEY]: typename,
-            ...mockData,
-          };
-        } else {
-          return mockData;
-        }
-      },
-    },
-    Field: {
-      /**
-       * Generate mock data
-       */
-      enter(fieldNode) {
-        const type = typeInfo.getType();
-        invariant(
-          type,
-          `Expected field to have a type: ${JSON.stringify(fieldNode)}`,
-        );
-        const namedType = getNamedType(type);
-        if (isScalarType(namedType)) {
-          const fieldWithMockData: typeof fieldNode = {
-            ...fieldNode,
-            userMockData: mockScalar(
-              fieldNode,
-              namedType,
-              typeInfo.getParentType()!,
-              resolveValue,
-            ),
-          };
-          return fieldWithMockData;
-        } else if (isEnumType(namedType)) {
-          const fieldWithMockData: typeof fieldNode = {
-            ...fieldNode,
-            userMockData: namedType.getValues()[0].value,
-          };
-          return fieldWithMockData;
-        } else {
-          invariant(
-            fieldNode.selectionSet,
-            "Expected field with selection set",
-          );
-          const fieldWithMockData: typeof fieldNode = {
-            ...fieldNode,
-            userMockData: mockCompositeType(
-              fieldNode,
-              assertCompositeType(namedType),
-              typeInfo.getParentType()!,
-              resolveValue,
-            ),
-          };
-          return fieldWithMockData;
-        }
-      },
-      /**
-       * Merge default mock data and explicit user mock data and return it
-       */
-      leave(fieldNode) {
-        const mockData = fieldNode.userMockData;
-        const mockFieldName = fieldNode.alias?.value || fieldNode.name.value;
-
-        if (!fieldNode.selectionSet) {
-          return {
-            [mockFieldName]: mockData,
-          };
-        }
-
-        const defaultData = (fieldNode.selectionSet as unknown) as MockData;
-        const data: MockData = {};
-        if (mockData) {
-          // Only use fields from the mockData that were actually selected
-          Object.keys(defaultData).forEach((fieldName) => {
-            data[fieldName] = (mockData as MockData).hasOwnProperty(fieldName)
-              ? (mockData as MockData)[fieldName]
-              : defaultData[fieldName];
-          });
-        } else {
-          Object.assign(data, defaultData);
-        }
-
-        const type = typeInfo.getType();
-        const isList = isListType(getNullableType(type!));
-        return {
-          [mockFieldName]: isList ? [data] : data,
-        };
-      },
-    },
-  };
-  return visit(documentDefinitionNode, visitWithTypeInfo(typeInfo, visitor));
-}
-
-/**
- * Finds the first fragment spread on an object type and keeps that plus any subsequent
- * fragment spreads for the same object type.
- */
-function reduceToSingleObjectTypeSelection(
-  selectionSetNode: SelectionSetNode,
-  explicitTypename: string | undefined,
-  schema: GraphQLSchema,
-  allDocumentDefinitionNodes: ReadonlyArray<DefinitionNode>,
-  typeInfo: TypeInfo,
-): SelectionSetNode {
-  let reduceToObjectType: GraphQLObjectType | null = null;
-  const reducer = (type: Maybe<GraphQLOutputType>) => {
-    if (isObjectType(type)) {
+  // RelayMockPayloadGenerator will execute documents that have optional
+  // boolean variables that are not passed by the user, but are required
+  // inside the document for @include/@skip to work. We rewrite the
+  // document to replace these variable references with hardcoded values.
+  //
+  // TODO: Is this a bug in RelayMockPayloadGenerator?
+  const undefinedBooleanVariables: string[] = [];
+  getOperationAST(operation.request.node)!.variableDefinitions?.forEach(
+    (variableDefinition) => {
       if (
-        // Select the first type we come across...
-        reduceToObjectType === null &&
-        // ...but only if the name matches, when an explicit one is given.
-        (explicitTypename === undefined || type.name === explicitTypename)
+        variableDefinition.type.kind === "NamedType" &&
+        variableDefinition.type.name.value === "Boolean" &&
+        operation.request.variables[variableDefinition.variable.name.value] ===
+          undefined
       ) {
-        reduceToObjectType = type;
-      } else if (
-        // If no types matched the explicit name yet...
-        reduceToObjectType === null ||
-        // ...or if the type is not the selected type...
-        reduceToObjectType !== type
-      ) {
-        // ...remove this fragment.
+        undefinedBooleanVariables.push(variableDefinition.variable.name.value);
+      }
+    },
+  );
+
+  const document =
+    undefinedBooleanVariables.length === 0
+      ? operation.request.node
+      : rewriteConditionals(operation.request.node, undefinedBooleanVariables);
+
+  const abstractTypeSelections: Path[] = [];
+
+  const result = executeSync({
+    schema: operation.schema,
+    document: document,
+    variableValues: operation.request.variables,
+    rootValue: {},
+    fieldResolver: (source: InternalMockData, args, _context, info) => {
+      // FIXME: This should not assume a single selection
+      const fieldNode: FieldNode = info.fieldNodes[0];
+      const selectionName = fieldNode.alias?.value ?? fieldNode.name.value;
+
+      // Explicit null value
+      if (source[selectionName] === null) {
         return null;
       }
-    }
-    return false;
-  };
-  return visit(
-    selectionSetNode,
-    visitWithTypeInfo(typeInfo, {
-      InlineFragment() {
-        return reducer(typeInfo.getType());
-      },
-      FragmentSpread(fragmentSpreadNode) {
-        const fragmentDefinitionNode = findFragmentDefinitionNode(
-          allDocumentDefinitionNodes,
-          fragmentSpreadNode.name.value,
+
+      const namedReturnType = getNamedType(info.returnType);
+      if (isCompositeType(namedReturnType)) {
+        // TODO: This 'is list' logic is also done by the value resolver,
+        // so probably need to refactor this code to actually leverage that.
+        const generateValue = (userValue?: {}) => {
+          // Explicit null value
+          if (userValue === null) {
+            return null;
+          }
+          const result = {
+            ...userValue,
+            ...mockCompositeType(
+              mockResolvers,
+              namedReturnType,
+              resolveValue,
+              fieldNode,
+              info,
+              args,
+              operation,
+              abstractTypeSelections,
+            ),
+          };
+          return result;
+        };
+        if (isListType(info.returnType)) {
+          const value = source[selectionName];
+          const result = Array.isArray(value)
+            ? value.map(generateValue)
+            : [generateValue(value as {})];
+          return result;
+        } else {
+          return generateValue(source[selectionName] as {} | undefined);
+        }
+      } else if (isScalarType(namedReturnType)) {
+        if (source[selectionName] !== undefined) {
+          return source[selectionName];
+        }
+        const result = mockScalar(
+          fieldNode,
+          args,
+          namedReturnType,
+          info,
+          source.__abstractType || info.parentType,
+          resolveValue,
+          isListType(info.returnType),
         );
-        const type = schema.getType(
-          fragmentDefinitionNode.typeCondition.name.value,
-        );
-        return reducer(type as GraphQLOutputType);
-      },
-    }),
-  );
+        return result;
+      } else if (isEnumType(namedReturnType)) {
+        if (source[selectionName] !== undefined) {
+          const value = source[selectionName];
+          return Array.isArray(value)
+            ? value.map((e) => String(e).toUpperCase())
+            : String(value).toUpperCase();
+        }
+        const enumValues = namedReturnType.getValues().map((e) => e.name);
+        return isListType(info.returnType) ? enumValues : enumValues[0];
+      } else {
+        return null;
+      }
+    },
+  });
+  invariant(result?.data, "Expected to generate a payload");
+  if (result.errors) {
+    throw new Error(`RelayMockPayloadGenerator: ${result.errors.join(", ")}`);
+  }
+
+  // Replace typenames of abstract type selections that had no concrete selection
+  // with the default mock typename.
+  abstractTypeSelections.forEach((pathInstance) => {
+    const path = pathToArray(pathInstance);
+    const object = path.reduce((prev, key) => prev[key], result.data!);
+    object.__typename = DEFAULT_MOCK_TYPENAME;
+  });
+
+  return result as { data: MockData };
 }
 
-function findFragmentDefinitionNode(
-  allDocumentDefinitionNodes: ReadonlyArray<DefinitionNode>,
-  name: string,
+function mockCompositeType(
+  mockResolvers: MockResolvers | null,
+  namedReturnType: GraphQLNamedType,
+  resolveValue: ValueResolver,
+  fieldNode: FieldNode,
+  info: GraphQLResolveInfo,
+  args: { [argName: string]: any },
+  operation: OperationDescriptor<GraphQLSchema, DocumentNode>,
+  abstractTypeSelections: Path[],
 ) {
-  const fragmentDefinitionNode = allDocumentDefinitionNodes.find(
-    (def) => def.kind === "FragmentDefinition" && def.name.value === name,
-  ) as FragmentDefinitionNode | undefined;
+  // Get the concrete type selection, concrete type on an abstract type
+  // selection, or the abstract type selection.
+  let typename = namedReturnType.name;
+  let abstractTypeSelectionOnly = false;
+  if (isAbstractType(namedReturnType)) {
+    const possibleTypeNames = info.fieldNodes.reduce<string[]>(
+      (acc, fieldNode) => [
+        ...acc,
+        ...getPossibleConcreteTypeNamesFromAbstractTypeSelections(
+          operation.schema,
+          fieldNode.selectionSet!,
+          info.fragments,
+        ),
+      ],
+      [],
+    );
+    if (possibleTypeNames?.length) {
+      typename = possibleTypeNames[0];
+    } else {
+      abstractTypeSelectionOnly = true;
+    }
+  }
+
+  const result: InternalMockData = {
+    ...getDefaultValues(
+      mockResolvers,
+      // If a mock resolver is provided for the abstract type, use it.
+      mockResolvers && mockResolvers[namedReturnType.name]
+        ? namedReturnType.name
+        : typename,
+      resolveValue,
+      fieldNode,
+      info,
+      args,
+    ),
+  };
+
+  if (isAbstractType(namedReturnType) && !result[TYPENAME_KEY]) {
+    if (abstractTypeSelectionOnly) {
+      // When no concrete type selection exists, it means only interface
+      // fields are selected, so we can just use the first possible type.
+      //
+      // We keep a reference to the path so we can rewrite the __typename
+      // field in the output after execution to reflect that this not a
+      // selection on a concrete type.
+      const possibleType = operation.schema.getPossibleTypes(
+        namedReturnType,
+      )[0];
+      invariant(
+        possibleType,
+        "Expected interface %s to be implemented by at least one concrete type",
+        namedReturnType.name,
+      );
+      typename = possibleType.name;
+      abstractTypeSelections.push(info.path);
+    }
+    result.__typename = typename;
+    result.__abstractType = namedReturnType;
+  }
+
+  return result;
+}
+
+function getDefaultValues(
+  mockResolvers: MockResolvers | null,
+  typename: string,
+  resolveValue: ValueResolver,
+  fieldNode: FieldNode,
+  info: GraphQLResolveInfo,
+  args: { [argName: string]: any },
+) {
+  const defaultValues =
+    mockResolvers![typename] &&
+    (resolveValue(
+      typename,
+      {
+        parentType: null,
+        name: fieldNode.name.value,
+        alias: fieldNode.alias?.value || null,
+        path: pathToArray(info.path).filter(isString),
+        args: args,
+      },
+      // FIXME: This is disabled here because we're currently doing this work
+      // in the field resolver's isCompositeType check.
+      // isListType(info.returnType),
+      false,
+    ) as MockData | undefined);
   invariant(
-    fragmentDefinitionNode,
-    `Expected a fragment by name '${name}' to exist`,
+    defaultValues === undefined || typeof defaultValues === "object",
+    "Expected mock resolver to return an object",
   );
-  return fragmentDefinitionNode;
+  return defaultValues;
+}
+
+function isString(value: any): value is string {
+  return typeof value === "string";
+}
+
+function getPossibleConcreteTypeNamesFromAbstractTypeSelections(
+  schema: GraphQLSchema,
+  selectionSet: SelectionSetNode,
+  fragments: { [key: string]: FragmentDefinitionNode },
+) {
+  const result: string[] = [];
+  selectionSet.selections.forEach((selection) => {
+    const fragmentDefinition = getFragmentSelection(selection, fragments);
+    if (fragmentDefinition) {
+      const typeName = fragmentDefinition.typeCondition?.name.value;
+      if (
+        // If undefined, it means the selection is an inline fragment on the
+        // current type; i.e. on an abstract type.
+        typeName === undefined ||
+        isAbstractType(schema.getType(typeName))
+      ) {
+        result.push(
+          ...getPossibleConcreteTypeNamesFromAbstractTypeSelections(
+            schema,
+            fragmentDefinition.selectionSet,
+            fragments,
+          ),
+        );
+      } else {
+        result.push(typeName);
+      }
+    }
+  });
+  return result;
+}
+
+function getFragmentSelection(
+  selection: SelectionNode,
+  fragments: { [key: string]: FragmentDefinitionNode },
+) {
+  if (selection.kind === "FragmentSpread") {
+    const fragmentDoc = fragments[selection.name.value];
+    invariant(
+      fragmentDoc,
+      "Expected a fragment with name %s to exist",
+      selection.name.value,
+    );
+    return fragmentDoc;
+  } else if (selection.kind === "InlineFragment") {
+    return selection;
+  }
 }
 
 function mockScalar(
   fieldNode: FieldNode,
+  args: { [argName: string]: any },
   type: GraphQLScalarType,
+  info: GraphQLResolveInfo,
   parentType: GraphQLCompositeType,
   resolveValue: ValueResolver,
+  plural: boolean,
 ) {
   if (fieldNode.name.value === TYPENAME_KEY) {
     return isAbstractType(parentType) ? DEFAULT_MOCK_TYPENAME : parentType.name;
   }
-  const args =
-    fieldNode.arguments &&
-    fieldNode.arguments.reduce(
-      (acc, arg) => ({ ...acc, [arg.name.value]: arg.value }),
-      {},
-    );
   const context: MockResolverContext = {
     name: fieldNode.name.value,
     alias: fieldNode.alias?.value || null,
     args,
+    path: pathToArray(info.path).filter(isString),
     parentType: isAbstractType(parentType)
       ? DEFAULT_MOCK_TYPENAME
       : parentType.name,
   };
-  return resolveValue(type.name, context, false, undefined);
+  return resolveValue(type.name, context, plural, undefined);
 }
 
-function mockCompositeType(
-  fieldNode: FieldNode,
-  type: GraphQLCompositeType,
-  parentType: GraphQLCompositeType,
-  resolveValue: ValueResolver,
-): MockData | undefined {
-  const args =
-    fieldNode.arguments &&
-    fieldNode.arguments.reduce(
-      (acc, arg) => ({ ...acc, [arg.name.value]: arg.value }),
-      {},
-    );
-  const context: MockResolverContext = {
-    name: fieldNode.name.value,
-    alias: fieldNode.alias?.value || null,
-    args,
-    parentType: isAbstractType(parentType)
-      ? DEFAULT_MOCK_TYPENAME
-      : parentType.name,
-  };
-  const data = resolveValue(type.name, context, false, undefined) as MockData;
-  // TODO: This is what they do upstream, it doesn't smell right
-  return typeof data === "object" ? data : undefined;
+function rewriteConditionals(
+  document: DocumentNode,
+  undefinedBooleanVariables: string[],
+) {
+  return visit(document, {
+    Directive(node) {
+      // Default value for @include is set to false
+      if (
+        node.name.value === "include" &&
+        node.arguments?.[0].value.kind === "Variable" &&
+        undefinedBooleanVariables.includes(node.arguments[0].value.name.value)
+      ) {
+        return {
+          ...node,
+          arguments: [
+            {
+              ...node.arguments[0],
+              value: {
+                kind: "BooleanValue",
+                value: false,
+              },
+            },
+          ],
+        };
+      }
+      // Default value for @skip is set to true
+      if (
+        node.name.value === "skip" &&
+        node.arguments?.[0].value.kind === "Variable" &&
+        undefinedBooleanVariables.includes(node.arguments[0].value.name.value)
+      ) {
+        return {
+          ...node,
+          arguments: [
+            {
+              ...node.arguments[0],
+              value: {
+                kind: "BooleanValue",
+                value: true,
+              },
+            },
+          ],
+        };
+      }
+    },
+  });
 }
